Mark-corrections[7,9,15,19,27,31,33,37,41,58,59,62,]
"In general, smart contracts are programs"
"They automatically execute"
"Smart contracts are deployed to the blockchain and are not controlled by the user - they run how they were programmed. User accounts interact with the contract by submitting transactions which execute functions from the contract".
"Arweave has its own version of a smart contract system. SmartWeave is a data-centered protocol built on top of the Arweave blockchain. It is unique because it deploys a lazy evaluation method. In other traditional smart contract systems it is a nodes responsibility to validate the results of a smart contract. SmartWeave's lazy evaluation transfers the burden of computation transaction validation to the smart contract user".
"The SmartWeave approach, on the other hand, has several advantages:"
"In general, it works like this. Contract source code (simple JavaScript function) and its initial state are deployed to Arweave and posted as transactions. They must meet the requirements from the protocol (e.g. contains specific tags). Then, when a user interacts with a contract, all of the prior transactions are evaluated until the end of chain of valid transactions. Here is the exact explanation."
"> 1. Loads all the contract interaction transactions up to the requested block height."
"> 3. Applies the sorted interactions onto the contract's handler function - evaluating the contract state up to the requested block height".
"In order to fully understand the process of contract creation we advise you to read the full protocol linked below".
"You can read more about smart contracts on Arweave in the following resources:"
"RedStone SmartContracts SDK is a new implementation of SmartWeave protocol that has been written from scratch. It focuses on performance and modularity by adding some enhancements to the first version of SmartWeave SDK".
"1. **The Core Protocol layer** - is the implementation of the original SmartWeave protocol and is responsible for communication with all of the SmartWeave smart contracts deployed on Arweave".
"2. **The Caching layer** - is built on top of the Core Protocol layer and allows the seperate caching results of each of the Core Protocol modules".
"Modular implementation has multiple advantages - principally, it can be tested and developed separately. The SmartWeave client is more customizable as different types of caches can be used for web and node environments. It also enables adding different loaders responsible for loading interaction transactions registered for a given contract. You can use `ArweaveGatewayInteractionsLoader` which uses Arweave gateway. Alternatively, it is possible to make use of `RedstoneGatewayInteractionsLoader` which uses a gateway dedicated to loading smart contracts transactions, available under [https://gateway.redstone.finance](https://gateway.redstone.finance/) url. You can view the instructions of how to use it in the SDK [here](https://github.com/redstone-finance/redstone-smartcontracts#using-the-redstone-gateway)".