Mark-corrections[5,9,17,20,]
"Profit sharing tokens (PST) are built using SmartWeave and are responsible for sharing the proceeds of any resources on Arweave. That means PSTs help monetize apps or websites built on Arweave. It's very simple. Each time a user interacts with an app, a specific amount of ARs is sent to PST holders based on their share. So if a holder owns 50% of the tokens, he will receive 50% of the revenue from the application".
"The amount of PST which will be transfered by the user while interacting with the app is specified by the developers creating the app. It can be everything from little payments up to large amounts for significant interactions. It can be fixed amount or a percentage of the interaction's value".
"A very interesting usage of PST is the ownership representation of the revenue from specific items. Imagine a musician producing new album. Users can then buy portions of revenue generated by this album. After the musician uploads this album to special platforms on Arweave (Bandplay, ArcLight), PST owners get a revenue split based on how much PST they own".
"You can find a PST guide in the [SmartWeave docs](https://github.com/ArweaveTeam/SmartWeave/blob/master/CREATE-PST.md)".