Mark-corrections[3,4,12]
"Now that we have a basic understanding of what PSTs are, we can dive into the code".
"We have prepared for you some basic setups consiting of Vue2 and Bootstrap for the frontend part, Typescript, Jest and ts-jest for tests, and ts-node for running scripts. You can of course use other tools/frameworks. However, we recommend using the ones we have prepared while going through this tutorial. You will need to write some code during the tutorial so the application is able to work. We advise you to read the instructions first and then try on your own. You can always view the final solution in [redstone-academy-pst/final](https://github.com/redstone-finance/redstone-academy/tree/main/redstone-academy-pst/final)".
"Install all dependencies connected with SmartWeave contracts. Do not worry, you will learn how to use them pretty soon. We will install specific versions of the libraries just so we are sure we are on the same page. You can of course use the most recent ones, but be aware that due to possible changes they may not be compatible with the tutorial".