Mark-corrections[43,89,121,167,208,212,214,245,]
"Time for an explanation".
"The above function will help us read the balance of the inidicated target address. It takes two arguments - contract computed state and destructured contract action which give us input to the interaction. Remember that we have three possible options to be returned from the interactions? In the above interaction we added two of them - thanks to simple error handling we can return `ContractError` or result".
"This one will help us minting some tokens to the caller's address. It takes two arguments - contract state and the destructured caller of the interaction as well as the input to the interaction. It adds tokens to the caller's address. It can throw `ContractError` exception or return contract's state".
"And the last one - the core function of our contract which will be responsible for transfering tokens between addresses. It takes two arguments - state and destructured caller of the interaction as well as the input to the interaction. It subtract's the indicated amount of tokens from caller's address and adds them to the target address. It can throw `ContractError` exception or return contract's state".
"`Handle` function is an asynchronous function and it returns a promise of type `ContractResult`. As mentioned above, it takes two arguments - state and action. It waits for one of the interactions to be called and returns the result of matching functions - the ones that we prepared earlier".
"Now comes the tricky part. We need to find a way to bundle our contract source so its output code is in Javascript and not typescript. We will use esbuild tool to achieve that result but of course you can use whichever bundler you'd like. We will not go into the details but you can view the bundling script here [https://github.com/redstone-finance/redstone-academy/blob/main/redstone-academy-pst/challenge/build.js](https://github.com/redstone-finance/redstone-academy/blob/main/redstone-academy-pst/challenge/build.js)".
"It takes the contract source file as an esbuild source file, bundles it and put's its compiled Javascript version in a `dist` folder."
"Now we just need to add a few commands to our `package.json` file that will simply remove everything from `dist` folder (which contains the minimized version of the source code), run the bundling script and additionally - copy `initial-state.json` file to the `dist` folder so we'll have all the files we need to deploy the contract in one place".