Mark-corrections[3,20,29,33,41,47]
"Before we start writing tests, we need to understand the different types of networks on Arweave blockchain".
"Instead of heading to arweave.net you can request and query nodes directly. Nodes are Arweave users with large storage spaces on their hard drives, they connect with the Arweave network and store data for others. You can use exactly the same API as for the gateway. A list of all currently running nodes can be found on [arweave.net/peers](https://arweave.net/peers)".
"Unlike mainnet, testnet means that the network is not running on its full capacity. It is used by developers to test different kind of actions and features before posting them to mainnet".
"1. **ArLocal** - is a local server which resembles Arweave mainnet. You need to run it locally on your computer".
"2. **RedStone testnet** - is a public testnet based on ArLocal. It does not require any instance to be run locally. The blocks are mined every 10s so you don't have to wait for the transaction as long as you would with a mainnet. You can also mine blocks manually, the same way you would do it on ArLocal".
"We will use ArLocal for our tests and RedStone public testnet for deploying contract and writing interactions. The main aim of the testnet is to a create close-to-real environment which is very similair to Arweave. There are of course small differences which we will point out".