Mark-corrections[5,7,22,78,94,101,105,121,132]
"As mentioned earlier, we will deploy our contract to the RedStone public testnet. You will notice that it is very similiar to how we deployed the contract to ArLocal. We just need to write a NodeJS script which will generate and fund the Arweave wallet, read, contract source and initial state files and deploy the contract to the testnet. At the end of this chapter, I'll show you how to repeat all these steps in order to deploy contract to Arweave mainnet."
"Like the last time - firstly, we will need to declare variables that will be needed in the script. Head to [challenge/src/tools/deploy-test-contract.ts](https://github.com/redstone-finance/redstone-academy/blob/main/redstone-academy-pst/challenge/src/tools/deploy-test-contract.ts) and add the following code:"
"Then, we need to initialize Arweave, this time we will point to the RedStone testnet. Write the following code in the asynchronous callback"
"And deploy contract using exactly the same method that is used in the tests. We will log the contract id to the console, so we can relate to it while writing interactions."
"All we need to do now is run the prepared script. We will use a typescript execution engine for NodeJS - ts-node. Run the following script in the root folder:"
"Remember that you need to bundle contract source file before deploying the contract. We've already done it in [**Writing PST contract** chapter](../writing-pst-contract/contract-source.md#-bundling-contract)."
"You've just deployed a contract to the RedStone testnet. You should see the contract id in the console output. You can also visit [SonAR](htttps://sonar.redstone.tools), switch to the testnet and search your contract :)"
"Now, the only thing you need to do is import the file with the wallet key and point to it when deploying your contract:"
"After deploying the contract to the mainnet you can also view it in SonAR. You just need to wait for the blocks to be mined. It can take up to 20 minutes."