Mark-corrections[3,30,58,68,74,76,]
"To run your application just execute the following command:"
"This time, we are using SDK's `SmartWeaveWebFactory` instead of `SmartWeaveNodeFactory` which will be safe to use in a web environment."
"You can get the state of your contract by destructuring the result of the PST method we already know - `currentState`. Thanks to having this object in the store, we will have constant access to the current state of the contract and we will be able to display all the addresses and their balances."
"If you want to create an application which enables interaction with the mainnet contract the only things you need to change in this section is indicating correct `host` in Arweave initialization (exactly like the last time - `arweave.net`) and pointing to the real wallet. This time it is best to detect if the user has ArConnect already installed. If yes - `arweaveWallet` object will be available in the `window` object and therefore you will be able to use ArConnect `connect` method. You can see its API in [ArConnect docs](https://docs.th8ta.org/arconnect/functions). You can also request necessary permissions. You can check a list of available permissions [here](https://docs.th8ta.org/arconnect/permissions).
"After you connect user wallet to the application you will need to use the `use_wallet` hook when initializing the contract. It will then connect the contract to the user's wallet. You can achieve this by writing following code:"