Mark-corrections[3,46,]
"We are displaying all the balances in the `BalancesList` component. We have access to the list thanks to `state` property we defined in the store in the Preparations section. Now all we need to do is find a way to properly mint some FC tokens to the wallet and transfer tokens between addresses."
"Finally, set the new balances list by calling `currentState` method:"