Mark-corrections[7,17,]
"The first 2 things you should think about before implementing a SmartWeave contract are:"
"Initially the `assets` object will be empty. But it will be filling with the newly generated assets and pointing to their owners. This state structure allows us to quickly understand who owns an asset or switch the ownership."